requirements.txt
anyio==3.5.0
async-generator==1.10
attrs==21.4.0
Authlib==0.15.5
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cycler==0.11.0
finnhub-python==2.4.13
fonttools==4.29.1
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
outcome==1.1.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
prompt-toolkit==3.0.28
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tda-api==1.5.1
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
ujson==5.1.0
urllib3==1.26.8
wcwidth==0.2.5
websocket-client==1.2.3
websockets==10.1
wsproto==1.0.0
